# Improve performance to_h method of Aws::Rails::SqsActiveJob::Configuration object