--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Improve to_h method's performance of Aws::Rails::SqsActiveJob::Configuration object
+
 3.9.1 (2023-12-19)
 ------------------