@roharon 2024-01-06
> Thanks. This needs a change log entry.

Then, do I need to update CHANGELOG?