@mullermp 2024-01-08
Yes, please update that file to add an issue line so that our automated system can release it.