@roharon 2024-01-09
@mullermp Thanks for your kindly guide. I changed it https://github.com/aws/aws-sdk-rails/pull/108/commits/aaaf9ab4178770b91a736a8ddab52ea0d4f44982